Move "create config" function to its own file in utils
